socialLogin: split ProtectedResourceMetadata code
Move OAuthProtectedResourceMetadataService to a shared project and
create an abstract superclass, in preparation for having another
implementation for the socialLogin feature.
