+++ dev/com.ibm.ws.security.openidconnect.client/test/com/ibm/ws/security/openidconnect/client/internal/OAuthProtectedResourceMetadataResolverTest.java
@@ -10,7 +10,6 @@
 package com.ibm.ws.security.openidconnect.client.internal;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 
@@ -165,149 +164,6 @@ public class OAuthProtectedResourceMetadataResolverTest {
         TestOidcClientConfig config = new TestOidcClientConfig("https://issuer.example.com");
 
         assertEquals("https://issuer.example.com", resolver.getAuthorizationServer(config));
-    }
-
-    // ---- ProtectedResourceRequestWrapper ------------------------------------
-
-    @Test
-    public void wrapperGetRequestURIReturnsProtectedResourcePath() {
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertEquals("/myApp/protected", wrapper.getRequestURI());
-    }
-
-    @Test
-    public void wrapperGetRequestURLStripsWellKnownPrefix() {
-        mock.checking(new Expectations() {
-            {
-                oneOf(mockRequest).getRequestURL();
-                will(returnValue(new StringBuffer("https://localhost:9443/.well-known/oauth-protected-resource/myApp/protected")));
-            }
-        });
-
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertEquals("https://localhost:9443/myApp/protected", wrapper.getRequestURL().toString());
-    }
-
-    @Test
-    public void wrapperGetRequestURLReturnsOriginalWhenNoWellKnownPrefix() {
-        mock.checking(new Expectations() {
-            {
-                oneOf(mockRequest).getRequestURL();
-                will(returnValue(new StringBuffer("https://localhost:9443/myApp/protected")));
-            }
-        });
-
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertEquals("https://localhost:9443/myApp/protected", wrapper.getRequestURL().toString());
-    }
-
-    @Test
-    public void wrapperGetServletPathReturnsProtectedResourcePath() {
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertEquals("/myApp/protected", wrapper.getServletPath());
-    }
-
-    @Test
-    public void wrapperGetContextPathReturnsEmpty() {
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertEquals("", wrapper.getContextPath());
-    }
-
-    @Test
-    public void wrapperGetPathInfoReturnsNull() {
-        OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper wrapper = new OAuthProtectedResourceMetadataResolver.ProtectedResourceRequestWrapper(mockRequest, "/myApp/protected");
-
-        assertNull(wrapper.getPathInfo());
-    }
-
-    // ---- createSignedMetadata null/blank guard ------------------------------
-
-    @Test
-    public void createSignedMetadataReturnsNullWhenBuilderIdIsNull() {
-        OAuthProtectedResourceMetadataResolver resolver = new OAuthProtectedResourceMetadataResolver();
-
-        String result = resolver.createSignedMetadata(null, null);
-
-        assertNull("Expected null when jwtBuilderRef is null", result);
-    }
-
-    @Test
-    public void createSignedMetadataReturnsNullWhenBuilderIdIsBlank() {
-        OAuthProtectedResourceMetadataResolver resolver = new OAuthProtectedResourceMetadataResolver();
-
-        String result = resolver.createSignedMetadata("  ", null);
-
-        assertNull("Expected null when jwtBuilderRef is blank", result);
-    }
-
-    // ---- createMetadataJson (existing) --------------------------------------
-
-    @Test
-    public void createMetadataJsonUsesIssuerIdentifierWhenPresent() {
-        OAuthProtectedResourceMetadataResolver resolver = new OAuthProtectedResourceMetadataResolver();
-        TestOidcClientConfig config = new TestOidcClientConfig();
-
-        config.id = "client1";
-        config.issuerIdentifier = "https://issuer.example.com";
-
-        String metadata = resolver.createMetadataJson(config, "https://localhost:9443/mcp", mockRequest);
-
-        assertEquals("{\"resource\":\"https:\\/\\/localhost:9443\\/mcp\",\"authorization_servers\":[\"https:\\/\\/issuer.example.com\"]}", metadata);
-    }
-
-    @Test
-    public void createMetadataJsonOmitsAuthorizationServersWhenIssuerIsMissing() throws Exception {
-        OAuthProtectedResourceMetadataResolver resolver = new OAuthProtectedResourceMetadataResolver();
-
-        TestOidcClientConfig config = new TestOidcClientConfig(null);
-        config.id = "client1";
-
-        String metadataJson = resolver.createMetadataJson(config, "https://localhost:9443/myApp/protected", mockRequest);
-
-        JSONObject metadata = JSONObject.parse(metadataJson);
-
-        assertEquals("https://localhost:9443/myApp/protected", metadata.get("resource"));
-        assertFalse(metadata.containsKey("authorization_servers"));
-    }
-
-    // ---- createMetadataJson with jwtBuilderRef ------------------------------
-
-    @Test
-    public void createMetadataJsonIncludesSignedMetadataWhenJwtBuilderRefIsConfigured() throws Exception {
-        TestResolver resolver = new TestResolver();
-        TestOidcClientConfig config = new TestOidcClientConfig("https://issuer.example.com", "myBuilder");
-        config.id = "client1";
-
-        String metadataJson = resolver.createMetadataJson(config, "https://localhost:9443/myApp/protected", mockRequest);
-
-        JSONObject metadata = JSONObject.parse(metadataJson);
-        assertEquals("test.signed.jwt", metadata.get("signed_metadata"));
-    }
-
-    @Test
-    public void createMetadataJsonOmitsSignedMetadataWhenJwtBuilderRefIsNull() throws Exception {
-        OAuthProtectedResourceMetadataResolver resolver = new OAuthProtectedResourceMetadataResolver();
-        TestOidcClientConfig config = new TestOidcClientConfig(null);
-        config.id = "client1";
-
-        String metadataJson = resolver.createMetadataJson(config, "https://localhost:9443/myApp/protected", mockRequest);
-
-        JSONObject metadata = JSONObject.parse(metadataJson);
-        assertFalse("signed_metadata should be absent when jwtBuilderRef is null", metadata.containsKey("signed_metadata"));
-    }
-
-    // ---- TestResolver -------------------------------------------------------
-
-    private static class TestResolver extends OAuthProtectedResourceMetadataResolver {
-        @Override
-        String createSignedMetadata(String jwtBuilderRef, JSONObject metadata) {
-            return "test.signed.jwt";
-        }
     }
 
     // ---- TestOidcClientConfig -----------------------------------------------
